Create PetTutor_motion_trigger_Arduino
This is a modification to the button triggered remote for the PT Feeder.  It is instead using the on-board IMU on the Arduino Nano 33 BLE to sense movement of the Arduino.  It is fairly stable but on occasion freezes up and needs to be reset to function again.  I believe there may be timing issues involved here.  Additionally, initial connection sometimes does not take without resetting the PT Feeder - I use PunchThrough BLE app to query the feeder, then exit the PT Feeder and the Arduino seems to connect OK.  I've noticed what appears to be a 4-attempt cycle to connect - the Arduino loops through the connection sequent about 4 times before it successfully connects.  This is based on the on-board LED flashing 3 times before finally connecting on the 4trh attempt.